moved files to registration location
