+++ frontend/pages/register/index.ts
